+++ modules/core/src/core.module.ts
@@ -3,7 +3,6 @@ import { NgModule, ModuleWithProviders } from "@angular/core";
 import { ReactiveFormsModule } from "@angular/forms";
 import { DynamicFormService } from "./service/dynamic-form.service";
 import { DynamicFormAutoFillService } from "./service/dynamic-form-autofill.service";
-import { DynamicFormRelationService } from "./service/dynamic-form-relation.service";
 import { DynamicIdDirective } from "./directive/dynamic-id.directive";
 import { DynamicTemplateDirective } from "./directive/dynamic-template.directive";
 
@@ -24,8 +23,7 @@ export class DynamicFormsCoreModule {
 
             providers: [
                 DynamicFormService,
-                DynamicFormAutoFillService,
-                DynamicFormRelationService
+                DynamicFormAutoFillService
             ]
         };
     }