+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 1.4.14
+
+### **Bugfixes**
+
+* Material template bug fixed 
+
+### **Features** 
+
+* major internal component refactoring
+
+
 # 1.4.13
 
 ### **Bugfixes**